// src/types/Anilist/Airing-sched.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export interface AiringSched {
    currentPage: number;
    hasNextPage: boolean;
    results:     Result[];
}

export interface Result {
    id:          string;
    malId:       number;
    episode:     number;
    airingAt:    number;
    title:       Title;
    country:     string;
    image:       string;
    description: null | string;
    cover:       string;
    genres:      string[];
    color:       string;
    rating:      number | null;
    releaseDate: number | null;
    type:        string;
}

export interface Title {
    romaji:        string;
    english:       null | string;
    native:        string;
    userPreferred: string;
}