// src/types/Anilist/Random.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export interface Random {
    id:              string;
    title:           Title;
    malId:           number;
    synonyms:        string[];
    isLicensed:      boolean;
    isAdult:         boolean;
    countryOfOrigin: string;
    image:           string;
    popularity:      number;
    cover:           string;
    description:     string;
    status:          string;
    releaseDate:     number;
    startDate:       EndDateClass;
    endDate:         EndDateClass;
    totalEpisodes:   number;
    currentEpisode:  number;
    rating:          number;
    duration:        number;
    genres:          string[];
    season:          string;
    studios:         string[];
    subOrDub:        string;
    type:            string;
    recommendations: Recommendation[];
    characters:      Character[];
    relations:       Relation[];
    episodes:        Episode[];
}

export interface Character {
    id:          number;
    role:        string;
    name:        Name;
    image:       string;
    voiceActors: VoiceActor[];
}

export interface Name {
    first:         string;
    last:          null | string;
    full:          string;
    native:        string;
    userPreferred: string;
}

export interface VoiceActor {
    id:       number;
    language: string;
    name:     Name;
    image:    string;
}

export interface EndDateClass {
    year:  number;
    month: number;
    day:   number;
}

export interface Episode {
    id:          string;
    title:       null;
    image:       string;
    number:      number;
    createdAt:   null;
    description: null;
    url:         string;
}

export interface Recommendation {
    id:       number;
    malId:    number;
    title:    Title;
    status:   string;
    episodes: number;
    image:    string;
    cover:    string;
    rating:   number;
    type:     string;
}

export interface Title {
    romaji:         string;
    english:        null | string;
    native:         string;
    userPreferred?: string;
}

export interface Relation {
    id:           number;
    relationType: string;
    malId:        number;
    title:        Title;
    status:       string;
    episodes:     null;
    image:        string;
    color:        string;
    type:         string;
    cover:        string;
    rating:       number;
}