// src/types/Anilist/Popular.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export interface Popular {
    currentPage: number;
    hasNextPage: boolean;
    results:     Result[];
}

export interface Result {
    id:            string;
    malId:         number;
    title:         Title;
    image:         string;
    trailer:       Trailer;
    description:   string;
    status:        Status;
    cover:         string;
    rating:        number | null;
    releaseDate:   number | null;
    color:         string;
    genres:        string[];
    totalEpisodes: number | null;
    duration:      number | null;
    type:          Type;
}

export enum Status {
    Completed = "Completed",
    NotYetAired = "Not yet aired",
    Ongoing = "Ongoing",
}

export interface Title {
    romaji:        string;
    english:       null | string;
    native:        string;
    userPreferred: string;
}

export interface Trailer {
    id?:        string;
    site?:      string;
    thumbnail?: string;
}

export enum Type {
    Tv = "TV",
}